Refactor App component: simplify YouTube player options management and remove unused video height logic
